pyproject.toml
[project]
name = "ygo-meta-ai"
version = "0.1.0"
description = "Yu-Gi-Oh meta AI: LLM game-playing agent + evolutionary Nash equilibrium deck builder"
requires-python = ">=3.9"

dependencies = [
    "httpx>=0.27.0",          # ygoinf HTTP client
    "anthropic>=0.28.0",      # Claude LLM API
    "nashpy>=0.0.17",         # Nash equilibrium solver
    "numpy>=1.26.0",
    "scipy>=1.13.0",          # LP fallback for large payoff matrices
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "anyio>=4.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
# Install these in a separate venv to avoid JAX version conflicts with modern numpy
ygoagent = [
    "jax<=0.4.28",
    "jaxlib<=0.4.28",
    "flax",
    "distrax",
    "chex",
    "uvicorn[standard]>=0.29.0",
    "fastapi>=0.111.0",
]

[project.scripts]
ygo-play = "ygo_meta.cli.play:main"
ygo-simulate = "ygo_meta.cli.simulate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ygo_meta"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
